+++ design/components-rn/vite.config.build.ts
@@ -1,7 +1,3 @@
 import customConfig from './vite.config.common.js';
 
-export default customConfig({
-  optimizeDeps: {
-    include: ['@expo/vector-icons'],
-  },
-});
+export default customConfig({});