+++ design/components-rn/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cloud-design/components-rn
 
+## 1.0.0
+
+### Minor Changes
+
+- change components rn eva icons to fontawesome icons
+
+### Patch Changes
+
+- Updated dependencies
+  - @cloud-dragon/react-utils@0.9.0
+  - @cloud-dragon/common-utils@0.9.0
+  - @cloud-dragon/world-countries@0.9.0
+
 ## 0.8.5
 
 ### Patch Changes